pyproject.toml
[tool.poetry]
name = "first-django"
version = "0.1.0"
description = ""
authors = ["jiyong.yoon <jyyoon0615@naver.com>"]
readme = "README.md"
packages = [{include = "first_django"}]

[tool.poetry.dependencies]
python = "^3.8"
arrow = "1.2.3"
asgiref = "3.7.2"
async-timeout = "4.0.2"
backports_zoneinfo = "0.2.1"
blinker = "1.6.2"
click = "8.1.6"
colorama = "0.4.6"
Django = "4.2.3"
django-rq = "2.8.1"
djangorestframework = "3.14.0"
Flask = "2.3.2"
importlib-metadata = "6.8.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.3"
python-dateutil = "2.8.2"
pytz = "2023.3"
redis = "5.0.0rc2"
rq = "1.15.1"
rq-dashboard = "0.6.1"
six = "1.16.0"
sqlparse = "0.4.4"
typing_extensions = "4.7.1"
tzdata = "2023.3"
Werkzeug = "2.3.6"
zipp = "3.16.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

